// src/api/sqlmodule/model/progressFinalizationResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { ProgressFinalizationResponseReason } from './progressFinalizationResponseReason';
import type { ProgressFinalizationResponseStatus } from './progressFinalizationResponseStatus';

/**
 * Результат завершения standalone- или platform-прохождения.
 */
export interface ProgressFinalizationResponse {
  progressId: string;
  status: ProgressFinalizationResponseStatus;
  bestScore: number;
  finalScore: number;
  isPassed: boolean;
  reason: ProgressFinalizationResponseReason;
  finalizedAt: string;
  canReturnToEducation: boolean;
  /** @nullable */
  returnUrl: string | null;
}