// src/api/sqlmodule/model/studentTaskProgressResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { StudentTaskProgressResponseFinalizationReason } from './studentTaskProgressResponseFinalizationReason';
import type { StudentTaskProgressResponseStatus } from './studentTaskProgressResponseStatus';

/**
 * Текущее прохождение SQL-задания.
 */
export interface StudentTaskProgressResponse {
  id: string;
  taskId: string;
  status: StudentTaskProgressResponseStatus;
  validationVersionId: string;
  bestScore: number;
  attemptsUsed: number;
  /** @nullable */
  attemptsRemaining: number | null;
  canSubmit: boolean;
  canFinalize: boolean;
  isPassed: boolean;
  /** @nullable */
  expiresAt: string | null;
  /** @nullable */
  finalScore: number | null;
  /** @nullable */
  finalizationReason: StudentTaskProgressResponseFinalizationReason;
  /** @nullable */
  finalizedAt: string | null;
}